SDK-1469: Rename SandboxClient to SandboxProfileClient
