+++ index.js
@@ -1,11 +1,11 @@
-const SandboxClientBuilder = require('./src/client.builder');
+const SandboxProfileClientBuilder = require('./src/profile/client.builder');
 const SandboxAttributeBuilder = require('./src/profile/request/attribute/attribute.builder');
 const SandboxAgeVerificationBuilder = require('./src/profile/request/attribute/derivation/age.verification.builder');
 const SandboxAnchorBuilder = require('./src/profile/request/attribute/anchor.builder');
 const TokenRequestBuilder = require('./src/profile/request/token.builder');
 
 module.exports = {
-  SandboxClientBuilder,
+  SandboxProfileClientBuilder,
   SandboxAttributeBuilder,
   SandboxAgeVerificationBuilder,
   SandboxAnchorBuilder,
